+++ packages/orm/CHANGELOG.md
@@ -1,5 +1,14 @@
 # orchid-orm
 
+## 1.6.28
+
+### Patch Changes
+
+- Remove undefined and void from json schema
+- Updated dependencies
+  - orchid-core@0.1.13
+  - pqb@0.10.24
+
 ## 1.6.27
 
 ### Patch Changes